+++ cmd/doctor.go
@@ -321,6 +321,7 @@ func checkPaths() error {
 	required := []string{
 		viper.GetString("archives_dir"),
 		viper.GetString("backups_dir"),
+		viper.GetString("cache_dir"),
 		viper.GetString("locks_dir"),
 		viper.GetString("overrides_dir"),
 		viper.GetString("tmp_dir"),